requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.2.0
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
billiard==4.2.0
black==23.9.1
celery==5.3.5
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.3.2
decorator==5.1.1
Django==4.2
django-jazzmin==2.6.0
django-leaflet==0.29.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
executing==2.0.0
factory-boy==3.3.0
Faker==19.13.0
frozenlist==1.4.0
flower==1.2.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
ipython==8.16.1
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
kombu==5.3.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.28.1
openapi-codec==1.3.2
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
ruff==0.1.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
vine==5.1.0
wcwidth==0.2.8
yarl==1.9.2
